split parser and asm gen and tokenizer into separate files, makefile doesnt work yet though
