+++ parser.c
@@ -1,0 +1,162 @@
+#include <token.h>
+
+
+static Node *expr(Token **rest, Token *token);
+static Node *equality(Token **rest, Token *token);
+static Node *relational(Token **rest, Token *token);
+static Node *add(Token **rest, Token *token);
+static Node *mul(Token **rest, Token *token);
+static Node *unary(Token **rest, Token *token);
+static Node *primary(Token **rest, Token *token);
+
+static Node *new_node(NodeType type) {
+  Node* node = calloc(1, sizeof(Node));
+  node->type = type;
+  return node;
+}
+
+static Node *new_binary(NodeType type, Node *left, Node *right) {
+  Node *node = new_node(type);
+  node->left = left;
+  node->right = right;
+  return node;
+}
+
+static Node *new_num(int val) {
+  Node *node = new_node(ND_NUM);
+  node->val = val;
+  return node;
+}
+
+static Node *new_unary(NodeType type, Node *expr) {
+  Node *node = new_node(type);
+  node->left = expr;
+  return node;
+}
+
+
+
+// expr = equality
+static Node *expr(Token **rest, Token *token) {
+  return equality(rest, token);
+}
+
+// equality = relational ("==" relational | "!=" relational)* 
+static Node *equality(Token **rest, Token *token) {
+  Node *node = relational(&token, token);
+  for (;;) {
+    if (equal(token, "==")) {
+      node = new_binary(ND_EQ, node, relational(&token, token->next));
+      continue;
+    }
+
+    if (equal(token, "!=")) {
+      node = new_binary(ND_NE, node, relational(&token, token->next));
+      continue;
+    }
+
+    *rest = token;
+    return node;
+  }
+}
+
+// relational = add ("<" add | "<=" add | ">" add | ">=" add)*
+static Node *relational(Token **rest, Token *token) {
+  Node *node = add(&token, token);
+
+  for (;;) {
+    if (equal(token, "<")) {
+      node = new_binary(ND_LT, node, add(&token, token->next));
+      continue;
+    }
+
+    if (equal(token, "<=")) {
+      node = new_binary(ND_LE, node, add(&token, token->next));
+      continue;
+    }
+
+    if (equal(token, ">")) {
+      node = new_binary(ND_LT, add(&token, token->next), node);
+      continue;
+    }
+
+    if (equal(token, ">=")) {
+      node = new_binary(ND_LE, add(&token, token->next), node);
+      continue;
+    }
+
+    *rest = token;
+    return node;
+  }
+}
+
+// add = mul ("+" mul | "-" mul)*
+static Node *add(Token **rest, Token *token) {
+  Node *node = mul(&token, token);
+  for (;;) {
+    if (equal(token, "+")) {
+      node = new_binary(ND_ADD, node, mul(&token, token->next));
+      continue;
+    }
+
+    if (equal(token, "-")) {
+      node = new_binary(ND_SUB, node, mul(&token, token->next));
+      continue;
+    }
+    *rest = token;
+    return node;
+  }
+}
+
+// mul = unary ("*" unary | "/" unary)*
+static Node *mul(Token **rest, Token *token) {
+  Node *node = unary(&token, token); 
+
+  for (;;) {
+    if (equal(token, "*")) {
+      node = new_binary(ND_MUL, node, unary(&token, token->next));
+      continue;
+    }
+
+    if (equal(token, "/")) {
+      node = new_binary(ND_DIV, node, unary(&token, token->next));
+      continue;
+    }
+
+    *rest = token;
+    return node;
+  }
+}
+
+// unary = ("+" | "-") unary 
+//       | primary
+static Node *unary(Token **rest, Token *token) {
+  if (equal(token, "+")) return unary(rest, token->next);
+  if (equal(token, "-")) return new_unary(ND_NEG, unary(rest, token->next));
+  return primary(rest, token);
+}
+
+// primary = "(" expr ")" | num
+static Node *primary(Token **rest, Token* token) {
+  // Skip parenthesis(idk how to spell it)
+  if (equal(token, "(")) {
+    Node *node = expr(&token, token->next);
+    *rest = skip(token, ")");
+    return node;
+  }
+
+  if (token->type == T_NUM) {
+    Node *node = new_num(token->val);
+    *rest = token->next;
+    return node;
+  } 
+  error_tok(token, "expected an expression");
+  return NULL; // It will exit before here
+}
+
+Node *parse(Token *token) {
+  Node *node = expr(&token, token);
+  if (token->kind != T_EOF)
+    error_tok(tok, "extra token");
+  return node;
+}
